refactor(MTRNIX-313): task 4 — rename storage.memory_freshness_pg to freshness_pg, add target_kind
